requirements.txt
numpy==2.4.4
opencv-python==4.13.0.92
ultralytics==8.4.38
Flask==3.1.3
Werkzeug==3.1.8
requests==2.33.1
python-dotenv==1.2.2
torch==2.11.0
torchvision==0.26.0
matplotlib==3.10.8
scipy==1.17.1
pillow==12.2.0
pyyaml==6.0.3
psutil==7.2.2